Revert back RoomList, Remove UniqueRoomList
