+++ src/main/java/seedu/address/model/ReadOnlyRoomList.java
@@ -6,9 +6,6 @@ import seedu.address.model.room.Room;
 
 public interface ReadOnlyRoomList {
 
-    /**
-     * Returns an unmodifiable view of the room list.
-     */
-    ObservableList<Room> getRoomList();
+    ObservableList<Room> getRoomObservableList();
 
 }
