+++ src/main/java/seedu/address/model/RoomList.java
@@ -1,89 +1,191 @@
 package seedu.address.model;
 
 import static java.util.Objects.requireNonNull;
+import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 
-import java.util.List;
+import java.util.Arrays;
+import java.util.Objects;
 import java.util.PriorityQueue;
+import java.util.logging.Logger;
 
+import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import seedu.address.commons.core.LogsCenter;
 import seedu.address.commons.core.index.Index;
 import seedu.address.model.room.Room;
-import seedu.address.model.room.UniqueRoomList;
+import seedu.address.model.room.exceptions.DuplicateRoomException;
+import seedu.address.model.room.exceptions.RoomNotFoundException;
 import seedu.address.model.task.Task;
+import seedu.address.storage.JsonPatientRecordsStorage;
 
 /**
- * Wraps all data at the app level
- * Duplicate patients are not allowed (by .isSamePatient comparison)
+ * Contains information regarding the Room information
  */
 public class RoomList implements ReadOnlyRoomList {
-    private final UniqueRoomList rooms;
-    /*
-     * The 'unusual' code block below is a non-static initialization block, sometimes used to avoid duplication
-     * between constructors. See https://docs.oracle.com/javase/tutorial/java/javaOO/initial.html
-     *
-     * Note that non-static init blocks are not recommended to use. There are other ways to avoid duplication
-     *   among constructors.
-     */
-    {
-        rooms = new UniqueRoomList();
-    }
+    private static final Logger logger = LogsCenter.getLogger(JsonPatientRecordsStorage.class);
 
+    private int numOfRooms;
+    private PriorityQueue<Room> rooms = new PriorityQueue<>();
+    private ObservableList<Room> internalList = FXCollections.observableArrayList();
+    private final ObservableList<Room> internalUnmodifiableList =
+            FXCollections.unmodifiableObservableList(internalList);
+
+    /** Creates default RoomList() object where all fields are null**/
     public RoomList() {}
 
     /**
-     * Creates a room record using the rooms in the {@code toBeCopied}
+     * Converts data from readOnlyRoomList to roomList
      */
-    public RoomList(ReadOnlyRoomList toBeCopied) {
+    public RoomList(ReadOnlyRoomList readOnlyRoomList) {
         this();
-        resetData(toBeCopied);
+        resetData(readOnlyRoomList);
     }
-
-    public int getNumOfRooms() {
-        return rooms.getNumOfRooms();
-    }
-
-    public void addTaskToRoom(Task task, Room room) {
-        rooms.addTaskToRoom(task, room);
-    }
-
     /**
-     * Replaces the contents of the room list with {@code rooms}.
+     * Creates a RoomList object using the information given in files containing information about
+     * which rooms are occupied and number of rooms
      */
-    public void setRooms(List<Room> rooms) {
-        this.rooms.setRooms(rooms);
+    public RoomList(PriorityQueue<Room> rooms, int numOfRooms) {
+        this.rooms = rooms;
+        this.numOfRooms = numOfRooms;
     }
 
-    public void setNumOfRooms(int numOfRooms) {
-        rooms.setNumOfRooms(numOfRooms);
+    public void resetData(ReadOnlyRoomList readOnlyRoomList) {
+        ObservableList<Room> roomLists = readOnlyRoomList.getRoomObservableList();
+        numOfRooms = roomLists.size();
+        rooms.addAll(roomLists);
+        internalList.addAll(roomLists);
     }
-
-    /// list overwrite operations
     /**
-     * Resets the existing data of this {@code RoomList} with {@code newData}.
+     * Returns Priority Queue of rooms
      */
-    public void resetData(ReadOnlyRoomList newData) {
-        requireNonNull(newData);
-        setRooms(newData.getRoomList());
-    }
-
-
     public PriorityQueue<Room> getRooms() {
-        return rooms.getRooms();
+        return this.rooms;
+    }
+
+    /**
+     * Returns number of rooms in hotel
+     */
+    public int getNumOfRooms() {
+        return numOfRooms;
+    }
+
+    public ObservableList<Room> getRoomObservableList() {
+        return internalList;
+    }
+
+    private void addRooms() {
+        if (numOfRooms < 0) {
+            return;
+        }
+        if (numOfRooms > internalList.size()) {
+            for (int i = internalList.size(); i < numOfRooms; i++) {
+                Room room = new Room(i + 1);
+                internalList.add(i, room);
+                rooms.add(room);
+            }
+        } else if (numOfRooms < internalList.size()) {
+            for (int i = numOfRooms; i < internalList.size(); i++) {
+                Room room = internalList.get(i);
+                rooms.remove(room);
+            }
+            int size = internalList.size();
+            for (int i = numOfRooms; i < size; i++) {
+                internalList.remove(numOfRooms);
+            }
+        }
+    }
+
+    /**
+     * Adds the number of the rooms in a hotel
+     *
+     * @param numOfRooms is the number of rooms to be added
+     */
+    public void addRooms(int numOfRooms) {
+        this.numOfRooms = numOfRooms;
+        addRooms();
+    }
+
+    /**
+     * Adds this room to the RoomList
+     * @param room is added to RoomList
+     */
+    public void addRooms(Room room) {
+        this.numOfRooms++;
+        rooms.add(room);
+        internalList.add(room);
+    }
+
+    /**
+     * Adds a task to a room.
+     *
+     * The room must exist in the {@code RoomList}.
+     *
+     * @param task The task to add.
+     * @param room The room to which the task should be added.
+     * @throws RoomNotFoundException if {@code room} is not in {@code RoomList}.
+     */
+    public void addTaskToRoom(Task task, Room room) {
+        requireAllNonNull(task, room);
+
+        int index = internalList.indexOf(room);
+        if (index == -1) {
+            throw new RoomNotFoundException();
+        }
+
+        room.addTask(task);
+        internalList.set(index, room);
+    }
+
+    /**
+     * Returns the backing list as an unmodifiable {@code ObservableList}.
+     */
+    public ObservableList<Room> asUnmodifiableObservableList() {
+        return internalUnmodifiableList;
     }
 
     @Override
-    public ObservableList<Room> getRoomList() {
-        return rooms.asUnmodifiableObservableList();
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+
+        RoomList roomList = (RoomList) o;
+        Room[] roomsForPQ = this.rooms.toArray(new Room[0]);
+        Room[] rooms1ForPQ = roomList.rooms.toArray(new Room[0]);
+
+        Room[] roomsForObservableList = internalList.toArray(new Room[0]);
+        Room[] rooms1FOrObservableList = roomList.internalList.toArray(new Room[0]);
+        return numOfRooms == roomList.numOfRooms
+                && Arrays.equals(roomsForPQ, rooms1ForPQ)
+                && Arrays.equals(roomsForObservableList, rooms1FOrObservableList);
     }
 
     /**
+     * Tests whether 2 PriorityQueues are equal by checking whether at each relative position they contain the equal
+     * rooms
+     */
+    public boolean equals(PriorityQueue<Room> rooms1, PriorityQueue<Room> rooms2) {
+        if (rooms1.size() != rooms2.size()) {
+            return false;
+        } else {
+            int size = rooms1.size();
+            for (int i = 0; i < size; i++) {
+                if (!rooms1.poll().equals(rooms2.poll())) {
+                    return false;
+                }
+            }
+            return true;
+        }
+    }
+    /**
      * Returns true if the list contains an equivalent room as the given argument.
-     * @param toCheck
-     * @return if app contains room or not
      */
     public boolean containsRoom(Room toCheck) {
         requireNonNull(toCheck);
-        return rooms.containsRoom(toCheck);
+        return internalList.stream().anyMatch(toCheck::isSameRoom);
     }
 
     /**
@@ -95,7 +197,31 @@ public class RoomList implements ReadOnlyRoomList {
      * @param editedRoom Room that has been changed.
      */
     public void setSingleRoom(Room target, Room editedRoom) {
-        rooms.setSingleRoom(target, editedRoom);
+        int index = internalList.indexOf(target);
+        if (index == -1) {
+            throw new RoomNotFoundException();
+        }
+
+        if (!target.isSameRoom(editedRoom) && containsRoom(editedRoom)) {
+            throw new DuplicateRoomException();
+        }
+        rooms.remove(target); // this and the next LOC is to replace the room in the priority queue
+        rooms.add(editedRoom);
+        internalList.set(index, editedRoom);
+    }
+    @Override
+    public int hashCode() {
+        int result = Objects.hash(numOfRooms, rooms, internalList);
+        result = 31 * result;
+        return result;
+    }
+
+    public void setNumOfRooms(int numOfRooms) {
+        this.numOfRooms = numOfRooms;
+    }
+
+    public void setRooms(PriorityQueue<Room> rooms) {
+        this.rooms = rooms;
     }
 
     /**
@@ -104,36 +230,15 @@ public class RoomList implements ReadOnlyRoomList {
      * @return Index Of room that is found.
      */
     public Index checkIfRoomPresent(Integer roomNumber) {
-        return rooms.checkIfRoomPresent(roomNumber);
-    }
-
-    /**
-     * Adds the number of the rooms in a hotel
-     *
-     * @param numOfRooms is the number of rooms to be added
-     */
-    public void addRooms(int numOfRooms) {
-        rooms.addRooms(numOfRooms);
-    }
-
-    /**
-     * Adds this room to the RoomList
-     * @param room is added to RoomList
-     */
-    public void addRooms(Room room) {
-        rooms.addRooms(room);
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        return other == this // short circuit if same object
-                || (other instanceof RoomList // instanceof handles nulls
-                && rooms.equals(((RoomList) other).rooms));
-    }
-
-    @Override
-    public int hashCode() {
-        return rooms.hashCode();
+        Index index = Index.fromZeroBased(0);
+        for (int i = 1; i <= internalList.size(); i++) {
+            int roomNum = internalList.get(i - 1).getRoomNumber();
+            boolean isValidRoom = (Integer.valueOf(roomNum)).equals(roomNumber);
+            if (isValidRoom) {
+                index = Index.fromZeroBased(i);
+                break;
+            }
+        }
+        return index;
     }
 }
-
