--- src/main/java/seedu/address/model/room/UniqueRoomList.java
@@ -1,238 +1,0 @@
-package seedu.address.model.room;
-
-import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
-
-import java.util.Arrays;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Objects;
-import java.util.PriorityQueue;
-
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import seedu.address.commons.core.index.Index;
-import seedu.address.model.room.exceptions.DuplicateRoomException;
-import seedu.address.model.room.exceptions.RoomNotFoundException;
-import seedu.address.model.task.Task;
-
-/**
- * A list of rooms that enforces uniqueness between its elements and does not allow nulls.
- *
- * Supports a minimal set of list operations.
- */
-public class UniqueRoomList implements Iterable<Room> {
-    private int numOfRooms;
-    private PriorityQueue<Room> rooms = new PriorityQueue<>();
-
-    private ObservableList<Room> internalList = FXCollections.observableArrayList();
-    private final ObservableList<Room> internalUnmodifiableList =
-            FXCollections.unmodifiableObservableList(internalList);
-
-    public void setNumOfRooms(int numOfRooms) {
-        this.numOfRooms = numOfRooms;
-    }
-
-    public void setRooms(PriorityQueue<Room> rooms) {
-        this.rooms = rooms;
-    }
-
-    /**
-     * Replaces the contents of this list with {@code rooms}.
-     */
-    public void setRooms(List<Room> rooms) {
-        requireAllNonNull(rooms);
-        internalList.setAll(rooms);
-        this.rooms = new PriorityQueue<>();
-        this.rooms.addAll(rooms);
-        this.numOfRooms = rooms.size();
-    }
-
-    //============================== Add Room operations =============================
-    private void addRooms() {
-        if (numOfRooms < 0) {
-            return;
-        }
-        if (numOfRooms > internalList.size()) {
-            for (int i = internalList.size(); i < numOfRooms; i++) {
-                Room room = new Room(i + 1);
-                internalList.add(i, room);
-                rooms.add(room);
-            }
-        } else if (numOfRooms < internalList.size()) {
-            for (int i = numOfRooms; i < internalList.size(); i++) {
-                Room room = internalList.get(i);
-                rooms.remove(room);
-            }
-            int size = internalList.size();
-            for (int i = numOfRooms; i < size; i++) {
-                internalList.remove(numOfRooms);
-            }
-        }
-    }
-
-    /**
-     * Adds the number of the rooms in a hotel
-     *
-     * @param numOfRooms is the number of rooms to be added
-     */
-    public void addRooms(int numOfRooms) {
-        this.numOfRooms = numOfRooms;
-        addRooms();
-    }
-
-    /**
-     * Adds this room to the RoomList
-     * @param room is added to RoomList
-     */
-    public void addRooms(Room room) {
-        this.numOfRooms++;
-        rooms.add(room);
-        internalList.add(room);
-    }
-
-    //============================== Edit Room operation =============================
-    /**
-     * Replaces the room {@code target} in the list with {@code editedRoom}.
-     * {@code target} must exist in the list.
-     * The room identity of {@code editedRoom} must not be the same as another existing room in the list.
-     *
-     * @param target Room to be changed.
-     * @param editedRoom Room that has been changed.
-     */
-    public void setSingleRoom(Room target, Room editedRoom) {
-        int index = internalList.indexOf(target);
-        if (index == -1) {
-            throw new RoomNotFoundException();
-        }
-
-        if (!target.isSameRoom(editedRoom) && containsRoom(editedRoom)) {
-            throw new DuplicateRoomException();
-        }
-        rooms.remove(target); // this and the next LOC is to replace the room in the priority queue
-        rooms.add(editedRoom);
-        internalList.set(index, editedRoom);
-    }
-
-    /**
-     * Adds a task to a room.
-     *
-     * The room must exist in the {@code RoomList}.
-     *
-     * @param task The task to add.
-     * @param room The room to which the task should be added.
-     * @throws RoomNotFoundException if {@code room} is not in {@code RoomList}.
-     */
-    //TODO Can this be done with setSingleRoom func?
-    public void addTaskToRoom(Task task, Room room) {
-        requireAllNonNull(task, room);
-
-        int index = internalList.indexOf(room);
-        if (index == -1) {
-            throw new RoomNotFoundException();
-        }
-
-        room.addTask(task);
-        internalList.set(index, room);
-    }
-
-
-    /**
-     * Returns Priority Queue of rooms
-     */
-    public PriorityQueue<Room> getRooms() {
-        return this.rooms;
-    }
-
-    /**
-     * Returns number of rooms in hotel
-     */
-    public int getNumOfRooms() {
-        return numOfRooms;
-    }
-
-
-    public ObservableList<Room> getRoomList() {
-        return internalList;
-    }
-
-    /**
-     * Returns the backing list as an unmodifiable {@code ObservableList}.
-     */
-    public ObservableList<Room> asUnmodifiableObservableList() {
-        return internalUnmodifiableList;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-
-        UniqueRoomList roomList = (UniqueRoomList) o;
-        Room[] roomsForPQ = this.rooms.toArray(new Room[0]);
-        Room[] rooms1ForPQ = roomList.rooms.toArray(new Room[0]);
-
-        Room[] roomsForObservableList = internalList.toArray(new Room[0]);
-        Room[] rooms1FOrObservableList = roomList.internalList.toArray(new Room[0]);
-        return numOfRooms == roomList.numOfRooms
-                && Arrays.equals(roomsForPQ, rooms1ForPQ)
-                && Arrays.equals(roomsForObservableList, rooms1FOrObservableList);
-    }
-
-    /**
-     * Tests whether 2 PriorityQueues are equal by checking whether at each relative position they contain the equal
-     * rooms
-     */
-    public boolean equals(PriorityQueue<Room> rooms1, PriorityQueue<Room> rooms2) {
-        if (rooms1.size() != rooms2.size()) {
-            return false;
-        } else {
-            int size = rooms1.size();
-            for (int i = 0; i < size; i++) {
-                if (!rooms1.poll().equals(rooms2.poll())) {
-                    return false;
-                }
-            }
-            return true;
-        }
-    }
-
-    public boolean containsRoom(Room toCheck) {
-        return internalList.stream().anyMatch(toCheck::isSameRoom);
-    }
-
-
-    @Override
-    public int hashCode() {
-        int result = Objects.hash(numOfRooms, rooms, internalList);
-        result = 31 * result;
-        return result;
-    }
-
-    /**
-     * Checks if the given room number is present in the application.
-     * @param roomNumber to check if it is in the application.
-     * @return Index Of room that is found.
-     */
-    public Index checkIfRoomPresent(Integer roomNumber) {
-        Index index = Index.fromZeroBased(0);
-        for (int i = 1; i <= internalList.size(); i++) {
-            int roomNum = internalList.get(i - 1).getRoomNumber();
-            boolean isValidRoom = (Integer.valueOf(roomNum)).equals(roomNumber);
-            if (isValidRoom) {
-                index = Index.fromZeroBased(i);
-                break;
-            }
-        }
-        return index;
-    }
-
-    @Override
-    public Iterator<Room> iterator() {
-        return internalList.iterator();
-    }
-}
-
